refactor: consolidate ObsEdit navigation logic
